+++ src/libtrx/game/console/internal.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void Console_History_Init(void);
+void Console_History_Shutdown(void);