+++ src/libtrx/include/libtrx/game/ui/widgets/prompt.h
@@ -7,6 +7,7 @@ void UI_Prompt_SetSize(UI_WIDGET *widget, int32_t width, int32_t height);
 
 void UI_Prompt_SetFocus(UI_WIDGET *widget, bool is_focused);
 void UI_Prompt_Clear(UI_WIDGET *widget);
+void UI_Prompt_ChangeText(UI_WIDGET *widget, const char *new_text);
 
 extern const char *UI_Prompt_GetPromptChar(void);
 extern int32_t UI_Prompt_GetCaretFlashRate(void);