+++ src/tr2/game/ui/common.c
@@ -16,6 +16,8 @@ UI_INPUT UI_TranslateInput(uint32_t system_keycode)
 {
     // clang-format off
     switch (system_keycode) {
+    case VK_UP:     return UI_KEY_UP;
+    case VK_DOWN:   return UI_KEY_DOWN;
     case VK_LEFT:   return UI_KEY_LEFT;
     case VK_RIGHT:  return UI_KEY_RIGHT;
     case VK_HOME:   return UI_KEY_HOME;